Transition from <input> elements to a keymap SVG
This pulls in a new dependency that implements tabs for us, and also changes the
keymap display from a sequence of `<input>` elements to an SVG that looks like
the actual keyboard.

Keycodes are not editable in-place, but keys can be selected, and the keycode
edited in the text input box below the keymap.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>