Improved detach handling
In the main app, listen to USB detach events, and if we find that a device with
our own VID/PID got detached, check `focus._port.isOpen` to see if it was us. If
it was, then display a warning, close the port forcibly, and go back to the
keyboard selection screen.

Fixes #31.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>