Flasher: Vastly improved robustness
Reset the keyboard ourselves, with higher timeouts than `AvrGirl`. This seems to
help with flashing failing at the post-reset stage, thus fixes #93.

We also go back to the keyboard selection page with a success message after a
successful flash, the `SaveChangesButton` thus does not have time to revert back
to active. This fixes #94.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>