KeyboardSelect: Provide a way to disconnect from the keyboard
We may want to disconnect from the keyboard without selecting another (there may
not even be another), but still keep Chrysalis running. For this reason, provide
a way to disconnect: when visiting the KeyboardSelect screen, if already
connected, and selecting the current device (which is the default state), the
"Connect" button will turn into "Disconnect", and act accordingly.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>