LayoutEditor: Update for the default/custom split
This is a major update to the LayoutEditor, which updates it to support the
default/custom split on the protocol level. While there, added the option to
hide default layers, and to set the `keymap.onlyCustom` flag too.

We also copy the default layers into EEPROM and toggle onlyCustom on if we see
that the EEPROM layers are all transparent.

Fixes #228.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>