KeyboardSelect: Support devices that share the same VID+PID
There are keyboards out there that share VID+PID with another. Instead of trying
to choose one, or autodetect which one we have, place both onto the list, and
let the end-user select the right one.

This requires an update to `@chrysalis-api/focus`, to a version that supports
this case too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>