Add a tool to update the screenshot
To make it easier to keep the screenshot up-to-date, added a small script that
takes a screenshot of a running Chrysalis, adds drop shadow around it, and saves
it as `data/screenshot.png`. Assumes ImageMagick is installed, and that we're
running under XOrg.

While there, also update the screenshot, using the new tool.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>