MainMenu: If there's a new version of Chrysalis, notify the user
If there's a new version of Chrysalis available (detected once on first open of
the sidebar), display an upgrade notice in the main menu, below the Chrysalis
version.

If our current version is a tagged build, we look for a newer tagged build on
GitHub. If it is an automatic snapshot build, we look for a newer snapshot on
S3.

Fixes #314.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>