Do not ask for confirmation when saving
We should only show a prompt when discarding unsaved changes, not when we're
explicitly saving them. To achieve this, `App.cancelContext()` now takes a
`dirty` argument, and only pops up the dialog when the flag is set. Cancelling
the context will set the dirty flag, any other call-site will not.

In practice, this fixes #335.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>