I'm not sure when this was broken but:
1. the default keymaps were being loaded using a different string than
what the file was actually named and
1. the `loadDefault` function set the data but never indicated that the
contents had changed. If there is supposed to be change detection it
isn't working (e.g. pasting manually doesn't work, but using the paste
from clipboard button does because it manually sets the isChange
property).

An additional improvement here would be to make the component actually detect
onChange on the TextField. For whatever reason the current onChange code
isn't functioning correctly. Maybe it detects correctly on keypress but
not paste? I'll leave that to someone else to figure out.

Signed-off-by: Murph Murphy <murph@clurictec.com>