tools/get-artifact: Verbosity and URL encoding fix
Added a bit of verbosity to the script, so we have a clue what it is doing at
the moment, and where the downloaded file will be saved. Also fixed an issue
with the script trying to download the wrong thing, because it forgot to escape
`+` in the version - now it does that.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>