flash: Better support for flashing with an external flasher
Sometimes the native JavaScript implementations aren't very reliable for one
reason or the other, so as an alternative, we now have a way to flash via an
external (but still bundled) application. This is currently available for AVR109
protocols, for any platform we ship an avrdude binary for.

One can switch between the two on the Preferences screen, in the Advanced
section, once a keyboard is connected. The setting will be remembered, and
defaults to off.

We also bundle avrdude (and its dlls) for the win32 platform, copied from
Arduino 1.8.12.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>