api/flash/Avr109: Close the serialport before flashing, if open
If we're connected to a keyboard that is already in bootloader mode, we need to
close the port first, before AvrGirl can connect to it. Thus, if our port is
open, close it. Without doing this, AvrGirl would not be able to connect to the
keyboard, and flashing would fail.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>