App: Always navigate to /keyboard-select at the end of usb detach
When handling an USB detach event, _always_ navigate to `/keyboard-select` at
the end, to avoid staying at a blank page. At the same time, always try to close
Focus, and also reset some of the application state.

While there, since we're in an async function, await all of those operations.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>