KeyboardSelect: Make it possible to install udev rules directly
Instead of displaying a cryptic error message when the device file is not
accessible, display a clearer one, and offer to install udev rules. After
installing the rules and triggering udev to run on the selected device file,
re-scan the devices.

If the installation succeeds, this will refresh the screen, and get rid of the
error. If the installation fails, we'll display a notification.

Fixes #534.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>