KeySelector: Update the groups that contain layer keys
We want to limit groups that contain layer keys to only show as many keys as we
have layers available. With the introduction of `MoveToLayer`, our groups
shifted, and we ended up putting OneShot Modifiers into this group, instead of
OSL, and completely missed `MoveToLayer`.

Updated the check, so it applies to the correct groups again.

Fixes #548.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>