github/workflows: Use the proper env vars for building
To build successfully, we need to pass `-std=c++14` to the c++ compiler while
compiling native dependencies. The easiest to achieve that is setting the
`CXXFLAGS` environment variable appropriately.

Similarly, we set `YARN_GPG` to no, to avoid timeouts on certain platforms.

Both of these were set for Travis, so we just copy them here, too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>