github/workflows: Update the build & publish workflow
This separates the build and publish steps, so we can have slightly different
publish steps depending on whether we're building a snapshot or a tagged
release. On a tagged release, we'll update the existing draft. On a snapshot,
we'll overwrite an existing pre-release.

Also update the document describing our release process.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>