Editor: Switch LayoutSelect to Autocomplete
We now use an Autocomplete component for the host layout selection, which gives
us a number of great benefits!

The most important is, that we can now type the name of the layout we want, and
the component will gradually narrow options down for us. With `<Select>`, only
the first letter was taken into account, which was unintuitive, to say the
least.

The dropdown is also smaller, which looks better - a list that covered most of
the screen just didn't feel good.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>