api/keymap: Load Croatian layouts from the osx set
We've been loading up all the windows layouts, under the assumption that it
covers all languages available via CLDR. It turns out, that's not the case.
Because naming is not 100% consistent, we can't load all the osx layouts and
filter out the duplicates, so we will be adding any missing languages as they
are reported.

With this patch, we lay the foundations for that, and load the Croatian layouts
from the osx set.

Fixes #647.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>