Editor: Rearrange the 104-key picker for a more standard layout
When we originally made the 104-key picker, it was displayed at the bottom of
the screen, and we wanted to save vertical space. For that reason, we ended up
putting the Fx keys to the right, and we did not include a few other keys, that
were offered elsewhere. Later on, we moved the picker from the bottom into a
popup in the center.

Being in a popup in the center, we no longer need to conserve vertical space,
nor do we need to exclude certain keys (which we no longer offer elsewhere!). As
such, rearrange the layout to a more standard one, with Esc and the Fx keys on
top, and some missing keys included.

Fixes #655.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>